VFolder: Add initial infrastructure
Using a m2m field between VirtualFolder and Unit will ease fast
retrieving of matching units and vfolders, also will allow to retrieve
units sorted by vfolder priority.

The priority field is a float in order to be able to set priorities
between 0 and 1. Validators are not run on instance save, so the we can't
use validators to ensure the priority is greater than zero.

Part of #3521.